requirements.txt
aiohttp==3.8.5 ; python_version >= "3.10" and python_version < "4"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4"
annotated-types==0.5.0 ; python_version >= "3.10" and python_version < "4"
astroid==2.15.6 ; python_version >= "3.10" and python_version < "4"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "4"
asyncio==3.4.3 ; python_version >= "3.10" and python_version < "4"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4"
black==23.9.1 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4"
click==8.1.7 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4" and (sys_platform == "win32" or platform_system == "Windows")
dill==0.3.7 ; python_version >= "3.10" and python_version < "4"
exceptiongroup==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
flake8==6.1.0 ; python_version >= "3.10" and python_version < "4"
frozenlist==1.4.0 ; python_version >= "3.10" and python_version < "4"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4"
isort==5.12.0 ; python_version >= "3.10" and python_version < "4"
lazy-object-proxy==1.9.0 ; python_version >= "3.10" and python_version < "4"
loguru==0.7.2 ; python_version >= "3.10" and python_version < "4"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4"
packaging==23.1 ; python_version >= "3.10" and python_version < "4"
pathspec==0.11.2 ; python_version >= "3.10" and python_version < "4"
platformdirs==3.10.0 ; python_version >= "3.10" and python_version < "4"
pluggy==1.3.0 ; python_version >= "3.10" and python_version < "4"
pycodestyle==2.11.0 ; python_version >= "3.10" and python_version < "4"
pydantic-core==2.6.3 ; python_version >= "3.10" and python_version < "4"
pydantic-settings==2.0.3 ; python_version >= "3.10" and python_version < "4"
pydantic==2.3.0 ; python_version >= "3.10" and python_version < "4"
pyflakes==3.1.0 ; python_version >= "3.10" and python_version < "4"
pylint==2.17.5 ; python_version >= "3.10" and python_version < "4"
pytest==7.4.2 ; python_version >= "3.10" and python_version < "4"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tomlkit==0.12.1 ; python_version >= "3.10" and python_version < "4"
typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "4"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "4" and sys_platform == "win32"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "4"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "4"
